# Conflicts:
#	doc/src/Commands_pair.rst
#	src/memory.h

+++ src/fix_bond_history.cpp
@@ -251,7 +251,7 @@ void FixBondHistory::post_neighbor()
 
 double FixBondHistory::memory_usage()
 {
-  return maxbond * ndata * sizeof(double);
+  return (double) maxbond * ndata * sizeof(double);
 }
 
 /* ---------------------------------------------------------------------- */
